% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxstream.R
\name{AIC.coxstream}
\alias{AIC.coxstream}
\title{Calculate the Akaike Information Criterion (AIC) for a 'coxstream' object}
\usage{
\method{AIC}{coxstream}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"coxstream"} for which the AIC is to
be calculated.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A numeric value representing the AIC of the model.
}
\description{
This function computes the AIC for a given 'coxstream' object. The AIC is a
measure of the relative quality of a statistical model for a given dataset.
It is calculated as: \eqn{-2 * logLik + 2 * k}, where \code{k} is the number
of parameters in the model.
}
\seealso{
\code{\link{logLik.coxstream}}
}
