% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basehaz.coxstream.R
\name{basehaz.coxstream}
\alias{basehaz.coxstream}
\title{Predict the baseline hazard function for \code{coxstream} objects}
\usage{
\method{basehaz}{coxstream}(object, newdata = NULL, conf.int = 0.95, ...)
}
\arguments{
\item{object}{An object of class \code{coxstream}.}

\item{newdata}{A numeric vector of time points at which to predict the
baseline hazard function. If \code{NULL}, the function will predict the
baseline hazard function at 100 equally spaced time points between the
boundaries of the \code{coxstream} object.}

\item{conf.int}{A numeric value between 0 and 1 indicating the confidence
level for the confidence interval. Default is 0.95.}

\item{...}{Additional arguments (not unused).}
}
\value{
A matrix with one row for each time point, and columns containing
the baseline hazard function, standard error, and confidence interval.
}
\description{
Predict the baseline hazard function for \code{coxstream} objects
}
