% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basehaz.hess.R
\name{basehaz.hess}
\alias{basehaz.hess}
\title{Hessian matrix of the baseline hazard function}
\usage{
basehaz.hess(t, y, parms)
}
\arguments{
\item{t}{A non-negative real number specifying the time point at which to
evaluate the baseline hazard function.}

\item{y}{the initial (state) values of the ODE system, which are not used.}

\item{parms}{A list of parameters containing the following components:
\describe{
\item{alpha}{The coefficients of the Bernstein polynomial.}
\item{degree}{A non-negative integer specifying the degree of the
Bernstein polynomial.}
\item{boundary}{Boundary knots at which to anchor the Bernstein polynomial}
}}
}
\value{
The Hessian matrix with respect to the coefficients of the baseline
hazard function evaluated at the time points.
}
\description{
Hessian matrix of the baseline hazard function
}
