% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basehaz.R
\name{basehaz}
\alias{basehaz}
\title{Baseline hazard function}
\usage{
basehaz(t, y, parms)
}
\arguments{
\item{t}{A non-negative real number specifying the time point at which to
evaluate the baseline hazard function.}

\item{y}{the initial (state) values of the ODE system, which are not used.}

\item{parms}{A list of parameters containing the following components:
\describe{
\item{alpha}{The coefficients of the Bernstein polynomial.}
\item{degree}{A non-negative integer specifying the degree of the
Bernstein polynomial.}
\item{boundary}{Boundary knots at which to anchor the Bernstein polynomial}
}}
}
\value{
The baseline hazard function evaluated at the specified time points.
}
\description{
Baseline hazard function
}
