% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basehaz_grad.R
\name{basehaz_grad}
\alias{basehaz_grad}
\title{Gradient of the Baseline Hazard Function}
\usage{
basehaz_grad(t, y, parms)
}
\arguments{
\item{t}{A non-negative real number specifying the time point at which to
evaluate the gradient of the baseline hazard function.}

\item{y}{The initial (state) values of the ODE system, which are not used in
Cox proportional hazards models.}

\item{parms}{A list of parameters containing the following components:
\describe{
\item{alpha}{The coefficients of the Bernstein polynomial.}
\item{degree}{A non-negative integer specifying the degree of the
Bernstein polynomial.}
\item{boundary}{Boundary knots at which to anchor the Bernstein
polynomial.}
}}
}
\value{
The gradient with respect to the coefficients of the baseline hazard
function evaluated at the specified time point.
}
\description{
Gradient of the Baseline Hazard Function
}
