% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{basehaz_ode}
\alias{basehaz_ode}
\title{Baseline Hazard Function for ODE Evaluation}
\usage{
basehaz_ode(t, y, parms)
}
\arguments{
\item{t}{A non-negative real number specifying the time point at which to
evaluate the baseline hazard function.}

\item{y}{The initial (state) values of the ODE system, which are not used in
Cox proportional hazards models.}

\item{parms}{A list of parameters containing the following components:
\describe{
\item{alpha}{The coefficients of the Bernstein polynomial.}
\item{n_basis}{A positive odd integer specifying the number of fourier
basis functions.}
\item{period}{The period of the Fourier basis functions.}
}}
}
\value{
The baseline hazard function evaluated at the specified time point.
}
\description{
Baseline Hazard Function for ODE Evaluation
}
