% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{chebyshev}
\alias{chebyshev}
\title{Compute Chebyshev Polynomial Basis}
\usage{
chebyshev(x, n_basis, boundary)
}
\arguments{
\item{x}{A numeric vector of values at which to evaluate the Chebyshev
polynomials.}

\item{n_basis}{An integer specifying the number of Chebyshev basis functions
(i.e., polynomial degrees) to compute, starting from degree 0 up to degree
n_basis - 1.}

\item{boundary}{A numeric vector of length 2 that defines the lower and upper
bounds for normalization.}
}
\value{
A matrix where each column corresponds to the Chebyshev polynomial of
a specific degree evaluated at the normalized input values.
}
\description{
This function generates a matrix of Chebyshev polynomial evaluations for a
given numeric vector, normalizing the input values to the interval
\eqn{[-1, 1]}, based on the provided boundary.
}
