% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.coxstream.R
\name{coef.coxstream}
\alias{coef.coxstream}
\title{Extract the coefficients from a \code{coxstream} object}
\usage{
\method{coef}{coxstream}(object, ...)
}
\arguments{
\item{object}{An object of class \code{coxstream}.}

\item{...}{Additional arguments (not used).}
}
\value{
A numeric vector of coefficients.
}
\description{
Extract the coefficients from a \code{coxstream} object
}
