% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoxStream-package.R
\docType{package}
\name{CoxStream-package}
\alias{CoxStream}
\alias{CoxStream-package}
\title{CoxStream: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SignorinoY/CoxStream}
  \item \url{http://gongziyang.com/CoxStream/}
  \item Report bugs at \url{https://github.com/SignorinoY/CoxStream/issues}
}

}
\author{
\strong{Maintainer}: Ziyang Gong \email{gongziyang@smail.swufe.edu.cn}

}
\keyword{internal}
