% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxstream.R
\name{cv.coxstream}
\alias{cv.coxstream}
\title{Cross-validation for \code{coxstream} objects}
\usage{
cv.coxstream(
  formula,
  data,
  boundary,
  idx_col,
  nu = 0.2,
  n_folds = 10,
  n_alphas = 5,
  n_basis_min = 3,
  seed = 0
)
}
\arguments{
\item{formula}{A formula expression as for regression models, of the form
\code{response ~ predictors}. The response must be a survival object as
returned by the \code{\link{Surv}} function.}

\item{data}{A data frame containing the variables in the model.}

\item{boundary}{A vector of length 2 containing the boundary knots.}

\item{idx_col}{The column name of the index column, which is used to
distinguish different patients.}

\item{nu}{A numeric value used to calculate the number of basis functions.
Default is 0.2.}

\item{n_folds}{A positive integer specifying the number of folds for
cross-validation. Default is 10.}

\item{n_alphas}{A positive integer specifying the number of alpha values to
test. Default is 5}

\item{n_basis_min}{A positive integer specifying the minimum number of
basis functions. Default is 3.}

\item{seed}{An integer specifying the random seed for reproducibility.
Default is 0.}
}
\value{
An object of class \code{cv.coxstream} is returned.
}
\description{
Cross-validation for \code{coxstream} objects
}
