% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxstream.R
\name{logLik.coxstream}
\alias{logLik.coxstream}
\title{Compute Log-Likelihood for 'coxstream' Objects}
\usage{
\method{logLik}{coxstream}(object, ...)
}
\arguments{
\item{object}{An object of class 'coxstream'.}

\item{...}{Additional arguments (currently unused).}
}
\value{
The log-likelihood value stored in the 'coxstream' object.
}
\description{
This function retrieves the log-likelihood value from an object of class
'coxstream'.
}
