% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_odd}
\alias{make_odd}
\title{Return odd number}
\usage{
make_odd(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector with all elements as odd numbers.
}
\description{
If the number is even, adds 1 to convert it into an odd number.
}
