% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{objective}
\alias{objective}
\title{Objective Function for Optimization}
\usage{
objective(
  par,
  time,
  status,
  x,
  subject,
  boundary,
  theta_prev,
  hess_prev,
  time_int
)
}
\arguments{
\item{par}{Numeric vector. Parameters to be optimized, including the
Bernstein polynomial coefficients of dimension \code{n_basis} (alpha)
and regression coefficients (beta).}

\item{time}{Numeric vector. Observed event or censoring times.}

\item{status}{Numeric vector. Event indicators (1 = event, 0 = censored).}

\item{x}{Numeric matrix. Covariate data for each observation.}

\item{subject}{Vector. Subject identifiers for repeated measurements.}

\item{boundary}{Numeric vector of length 2. The lower and upper bounds for
the Bernstein polynomial basis functions.}

\item{theta_prev}{Numeric vector. Previous parameter estimates used in the
proximal term, including the Bernstein polynomial coefficients of dimension
\code{n_basis_prev} (alpha) and regression coefficients (beta). The number
of basis functions in the previous estimate must be greater than or equal
to the number of basis functions in the current estimate.}

\item{hess_prev}{Numeric matrix. Previous Hessian matrix used in the proximal
term, which is a square matrix of size \code{length(theta_prev)}.}

\item{time_int}{Numeric vector. Additional time points stored in a named
vector, where the names correspond to the subject identifiers. These time
points are used for evaluating the baseline hazard function and its
derivatives.}
}
\value{
The objective value with attributes "gradient" and "hessian".
}
\description{
Computes the objective value, gradient, and Hessian for the penalized
likelihood in the Cox model with a Bernstein polynomial bases baseline
hazard. This function is designed for use in optimization routines.
}
