% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.coxstream.R
\name{summary.coxstream}
\alias{summary.coxstream}
\title{Summary method for \code{coxstream} object}
\usage{
\method{summary}{coxstream}(object, conf.int = 0.95, ...)
}
\arguments{
\item{object}{An object of class \code{coxstream}.}

\item{conf.int}{A numeric value between 0 and 1 indicating the confidence
level for the confidence interval. Default is 0.95.}

\item{...}{Additional arguments (not unused).}
}
\value{
An object of class \code{"summary.coxstream"}, containing the following
components:
\describe{
\item{call}{The matched call.}
\item{coefficients}{A matrix with one row for each coefficient, and
columns containing the coefficient, the hazard ratio (exp(coef)), standard
error, Wald statistic, and P value.}
\item{conf.int}{A matrix with one row for each coefficient, containing the
confidence limits for exp(coef).}
}
}
\description{
Summary method for \code{coxstream} object
}
