% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcov.coxstream.R
\name{vcov.coxstream}
\alias{vcov.coxstream}
\title{Variance-covariance method for \code{coxstream} object}
\usage{
\method{vcov}{coxstream}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"coxstream"}.}

\item{...}{Unused.}
}
\value{
A matrix of the variance-covariance matrix of the coefficients.
}
\description{
Variance-covariance method for \code{coxstream} object
}
