% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{basehaz_grad}
\alias{basehaz_grad}
\title{Gradient of the Baseline Hazard Function}
\usage{
basehaz_grad(t, y, parms)
}
\arguments{
\item{t}{A non-negative real number specifying the time point at which to
evaluate the gradient of the baseline hazard function.}

\item{y}{The initial (state) values of the ODE system, which are not used in
Cox proportional hazards models.}

\item{parms}{A list of parameters containing the following components:
\describe{
\item{alpha}{The coefficients of the Chebyshev polynomial.}
\item{n_basis}{A positive odd integer specifying the number of Chebyshev
basis functions.}
\item{boundary}{A numeric vector of length 2 that defines the lower and
upper bounds for normalization.}
}}
}
\value{
The gradient with respect to the coefficients of the baseline hazard
function evaluated at the specified time point.
}
\description{
Gradient of the Baseline Hazard Function
}
