% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basehaz_hess.R
\name{basehaz_hess}
\alias{basehaz_hess}
\title{Hessian Matrix of the Baseline Hazard Function}
\usage{
basehaz_hess(t, y, parms)
}
\arguments{
\item{t}{A non-negative real number specifying the time point at which to
evaluate the Hessian matrix of the baseline hazard function.}

\item{y}{The initial (state) values of the ODE system, which are not used in
Cox proportional hazards models.}

\item{parms}{A list of parameters containing the following components:
\describe{
\item{alpha}{The coefficients of the Bernstein polynomial.}
\item{degree}{A non-negative integer specifying the degree of the
Bernstein polynomial.}
\item{boundary}{Boundary knots at which to anchor the Bernstein
polynomial.}
}}
}
\value{
The Hessian matrix with respect to the coefficients of the baseline
hazard function evaluated at the specified time point.
}
\description{
Hessian Matrix of the Baseline Hazard Function
}
