% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{basehaz_hess}
\alias{basehaz_hess}
\title{Hessian Matrix of the Baseline Hazard Function}
\usage{
basehaz_hess(t, y, parms)
}
\arguments{
\item{t}{A non-negative real number specifying the time point at which to
evaluate the Hessian matrix of the baseline hazard function.}

\item{y}{The initial (state) values of the ODE system, which are not used in
Cox proportional hazards models.}

\item{parms}{A list of parameters containing the following components:
\describe{
\item{alpha}{The coefficients of the bernstein polynomial.}
\item{n_basis}{A positive odd integer specifying the number of bernstein
basis functions.}
\item{boundary}{A numeric vector of length 2 that defines the lower and
upper bounds for normalization.}
}}
}
\value{
The Hessian matrix with respect to the coefficients of the baseline
hazard function evaluated at the specified time point.
}
\description{
Hessian Matrix of the Baseline Hazard Function
}
