% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxstream-package.R
\docType{package}
\name{coxstream-package}
\alias{coxstream-package}
\title{coxstream: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SignorinoY/coxstream}
  \item \url{http://gongziyang.com/coxstream/}
  \item Report bugs at \url{https://github.com/SignorinoY/coxstream/issues}
}

}
\author{
\strong{Maintainer}: Ziyang Gong \email{gongziyang@smail.swufe.edu.cn} (\href{https://orcid.org/0009-0001-4870-9102}{ORCID})

}
\keyword{internal}
