% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxstream.R
\name{coxstream}
\alias{coxstream}
\title{Streaming Cox model with repeated observations}
\usage{
coxstream(formula, data, degree, boundary, idx_col)
}
\arguments{
\item{formula}{A formula expression as for regression models, of the form
\code{response ~ predictors}. The response must be a survival object as
returned by the \code{\link{Surv}} function.}

\item{data}{A data frame containing the variables in the model.}

\item{degree}{The degree of the Bernstein polynomial.}

\item{boundary}{A vector of length 2 containing the boundary knots.}

\item{idx_col}{The column name of the index column, which is used to
distinguish different patients.}
}
\value{
An object of class \code{coxstream}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
library(coxstream)
formula <- survival::Surv(time, status) ~ X1 + X2 + X3 + X4 + X5
fit <- coxstream(
  formula, sim[sim$batch_id == 1, ],
  degree = 3, boundary = c(0, 3), idx_col = "patient_id"
)
for (batch in 2:10) {
  fit <- update(fit, sim[sim$batch_id == batch, ])
}
summary(fit)
}
