% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.coxstream.R
\name{update.coxstream}
\alias{update.coxstream}
\title{Update the \code{coxstream} with new data.}
\usage{
\method{update}{coxstream}(object, data, ...)
}
\arguments{
\item{object}{A \code{coxstream} object.}

\item{data}{A data frame containing the variables in the model.}

\item{...}{Additional arguments (not used).}
}
\value{
An object of class \code{coxstream}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
