% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxstream.R
\name{update.coxstream}
\alias{update.coxstream}
\title{Update the \code{coxstream} with new data.}
\usage{
\method{update}{coxstream}(object, data, n_basis = "auto", ...)
}
\arguments{
\item{object}{A \code{coxstream} object.}

\item{data}{A data frame containing the variables in the model.}

\item{n_basis}{Either a positive odd integer specifying the number of basis
functions, or the string "auto" to automatically determine this number based
\eqn{[\alpha N^{\nu}]}, where \eqn{N} is the number of unique survival times.}

\item{...}{Additional arguments (not used).}
}
\value{
An object of class \code{coxstream}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
