% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxstream.R
\name{vcov.coxstream}
\alias{vcov.coxstream}
\title{Extract variance-covariance matrix from a \code{coxstream} object}
\usage{
\method{vcov}{coxstream}(object, ...)
}
\arguments{
\item{object}{An object of class \code{coxstream}.}

\item{...}{Additional arguments (not unused).}
}
\value{
A matrix representing the variance-covariance matrix of the
coefficients.
}
\description{
Extract variance-covariance matrix from a \code{coxstream} object
}
